add listconfessions api
